checkers_board_class
